mock: moved mock code to its own crate
