+++ mock/Cargo.toml
@@ -11,13 +11,25 @@ keywords = []
 edition = "2021"
 rust-version = "1.67"
 
+[[bin]]
+name = "mock"
+path = "src/main.rs"
+required-features = ["executable"]
+
+[features]
+serde = ["dep:serde", "crypto/serde"]
+executable = ["serde"]
+
 [dependencies]
+assembly = { workspace = true }
 clap = { version = "4.4" }
+crypto = { workspace = true }
+env_logger = { version = "0.10" }
 getrandom = { version = "0.2", features = ["js"] }
+hex = "0.4"
+miden-objects = { path = "../objects" , features = ["mock", "serde", "log", "testing"]}
+postcard = { version = "1.0", features = [ "alloc" ] }
 rand = { version = "0.8" }
 rand_pcg = { version = "0.3", features = ["serde1"] }
-miden-objects = { path = "../objects" , features = ["mock", "serde", "log"]}
-crypto = { workspace = true }
-postcard = { version = "1.0", features = [ "alloc" ] }
-env_logger = { version = "0.10" }
-hex = "0.4"
+serde = { version = "1.0", optional = true, default-features = false, features = ["derive"] }
+vm-core = { workspace = true }
