+++ mock/src/builders/error.rs
@@ -1,0 +1,24 @@
+use core::fmt::Display;
+use crypto::merkle::MerkleError;
+use miden_objects::AccountError;
+
+#[derive(Debug)]
+pub enum AccountBuilderError {
+    AccountError(AccountError),
+    MerkleError(MerkleError),
+
+    /// When the created [AccountId] doesn't match the builder's configured [AccountType].
+    SeedAndAccountTypeMismatch,
+
+    /// When the created [AccountId] doesn't match the builder's `on_chain` config.
+    SeedAndOnChainMismatch,
+}
+
+impl Display for AccountBuilderError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for AccountBuilderError {}
