+++ mock/src/builders/nonfungible_asset.rs
@@ -1,0 +1,47 @@
+use miden_objects::{
+    assets::{NonFungibleAsset, NonFungibleAssetDetails},
+    AccountId, AccountType, AssetError,
+};
+use rand::{distributions::Standard, Rng};
+
+/// Builder for an `NonFungibleAssetDetails`, the builder can be configured and used multipled times.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+pub struct NonFungibleAssetDetailsBuilder<T> {
+    faucet_id: AccountId,
+    rng: T,
+}
+
+impl<T: Rng> NonFungibleAssetDetailsBuilder<T> {
+    pub fn new(faucet_id: AccountId, rng: T) -> Result<Self, AssetError> {
+        if !matches!(faucet_id.account_type(), AccountType::NonFungibleFaucet) {
+            return Err(AssetError::not_a_non_fungible_faucet_id(faucet_id));
+        }
+
+        Ok(Self { faucet_id, rng })
+    }
+
+    pub fn build(&mut self) -> Result<NonFungibleAssetDetails, AssetError> {
+        let data = (&mut self.rng).sample_iter(Standard).take(5).collect();
+        NonFungibleAssetDetails::new(self.faucet_id, data)
+    }
+}
+
+/// Builder for an `NonFungibleAsset`, the builder can be configured and used multipled times.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+pub struct NonFungibleAssetBuilder<T> {
+    details_builder: NonFungibleAssetDetailsBuilder<T>,
+}
+
+impl<T: Rng> NonFungibleAssetBuilder<T> {
+    pub fn new(faucet_id: AccountId, rng: T) -> Result<Self, AssetError> {
+        let details_builder = NonFungibleAssetDetailsBuilder::new(faucet_id, rng)?;
+        Ok(Self { details_builder })
+    }
+
+    pub fn build(&mut self) -> Result<NonFungibleAsset, AssetError> {
+        let details = self.details_builder.build()?;
+        NonFungibleAsset::new(&details)
+    }
+}
