+++ mock/src/chain.rs
@@ -1,0 +1,591 @@
+use super::builders::{
+    accountid_build_details, AccountBuilder, AccountIdBuilder, AccountStorageBuilder,
+    FungibleAssetBuilder, NonFungibleAssetBuilder,
+};
+use core::fmt;
+use crypto::{
+    hash::rpo::RpoDigest as Digest,
+    merkle::{NodeIndex, SimpleSmt, TieredSmt},
+    utils::collections::Vec,
+    Felt, FieldElement, StarkField, Word,
+};
+use miden_objects::{
+    assets::Asset,
+    notes::{Note, NoteInclusionProof, NOTE_LEAF_DEPTH, NOTE_TREE_DEPTH},
+    Account, AccountId, AccountType, BlockHeader, ChainMmr, StorageItem, DEFAULT_ACCOUNT_CODE,
+};
+use rand::{Rng, SeedableRng};
+
+/// Initial timestamp value
+const TIMESTAMP_START: Felt = Felt::new(1693348223);
+/// Timestep of timestamp on each new block
+const TIMESTAMP_STEP: Felt = Felt::new(10);
+
+#[derive(Default, Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+pub struct Objects<R> {
+    /// Holds the account and its corresponding seed.
+    accounts: Vec<(Account, Word)>,
+    fungible_faucets: Vec<(AccountId, FungibleAssetBuilder)>,
+    nonfungible_faucets: Vec<(AccountId, NonFungibleAssetBuilder<R>)>,
+    notes: Vec<Note>,
+    nullifiers: Vec<Digest>,
+}
+
+impl<R: Rng> Objects<R> {
+    pub fn new() -> Self {
+        Self {
+            accounts: vec![],
+            fungible_faucets: vec![],
+            nonfungible_faucets: vec![],
+            notes: vec![],
+            nullifiers: vec![],
+        }
+    }
+
+    /// Update this instance with objects inserted in the chain.
+    ///
+    /// This method expects `pending` to be a list of objects in the pending block, and for
+    /// this instance to be the set of objects added to the chain. Once the pending block is
+    /// sealed and the auxiliary data is produced (i.e. the notes tree), this method can be
+    /// called to 1. update the pending objects with the new data 2. move the objects to this
+    /// container.
+    pub fn update_with(
+        &mut self,
+        pending: &mut Objects<R>,
+        header: BlockHeader,
+        notes: &SimpleSmt,
+    ) {
+        self.accounts.append(&mut pending.accounts);
+        self.fungible_faucets.append(&mut pending.fungible_faucets);
+        self.nonfungible_faucets.append(&mut pending.nonfungible_faucets);
+
+        pending.set_notes_proofs(header, notes);
+        self.notes.append(&mut pending.notes);
+        pending.nullifiers.clear(); // nullifiers are saved in the nullifier TSTM
+    }
+
+    /// Creates a [SimpleSmt] tree from the `notes`.
+    ///
+    /// The root of the tree is a commitment to all notes created in the block. The commitment
+    /// is not for all fields of the [Note] struct, but only for note metadata + core fields of
+    /// a note (i.e., vault, inputs, script, and serial number).
+    pub fn build_notes_tree(&self) -> SimpleSmt {
+        let mut entries = Vec::with_capacity(self.notes.len() * 2);
+
+        entries.extend(self.notes.iter().enumerate().map(|(index, note)| {
+            let tree_index = (index * 2) as u64;
+            (tree_index, note.hash().into())
+        }));
+        entries.extend(self.notes.iter().enumerate().map(|(index, note)| {
+            let tree_index = (index * 2 + 1) as u64;
+            (tree_index, note.metadata().into())
+        }));
+
+        SimpleSmt::with_leaves(NOTE_LEAF_DEPTH, entries).unwrap()
+    }
+
+    /// Given the [BlockHeader] and its notedb's [SimpleSmt], set all the [Note]'s proof.
+    ///
+    /// Update the [Note]'s proof once the [BlockHeader] has been created.
+    fn set_notes_proofs(&mut self, header: BlockHeader, notes: &SimpleSmt) {
+        self.notes.iter_mut().enumerate().for_each(|(index, note)| {
+            let auth_index =
+                NodeIndex::new(NOTE_TREE_DEPTH, index as u64).expect("index bigger than 2**20");
+            let note_path =
+                notes.get_path(auth_index).expect("auth_index outside of SimpleSmt range");
+            note.set_proof(
+                NoteInclusionProof::new(
+                    header.block_num(),
+                    header.sub_hash(),
+                    header.note_root(),
+                    index as u64,
+                    note_path,
+                )
+                .expect("Invalid data provided to proof constructor"),
+            );
+        });
+    }
+}
+
+/// Structure chain data, used to build necessary openings and to construct [BlockHeader]s.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+pub struct MockChain<R> {
+    /// An append-only structure used to represent the history of blocks produced for this chain.
+    chain: ChainMmr,
+
+    /// History of produced blocks.
+    blocks: Vec<BlockHeader>,
+
+    /// Tree containing the latest `Nullifier`'s tree.
+    nullifiers: TieredSmt,
+
+    /// Tree containing the latest hash of each account.
+    // TODO: change this to a TieredSmt with 64bit keys.
+    accounts: SimpleSmt,
+
+    /// RNG used to seed builders.
+    ///
+    /// This is used to seed the [AccountBuilder] and the [NonFungibleAssetBuilder].
+    rng: R,
+
+    /// Builder for new [AccountId]s of faucets.
+    account_id_builder: AccountIdBuilder<R>,
+
+    /// Objects that have been created and commited to a block.
+    ///
+    /// These can be used to perform additional operations on a block.
+    ///
+    /// Note:
+    /// - The [Note]s in this container have the `proof` set.
+    objects: Objects<R>,
+
+    /// Objects that have been created and are waiting for a block.
+    ///
+    /// These objects will become available once the block is sealed.
+    ///
+    /// Note:
+    /// - The [Note]s in this container do not have the `proof` set.
+    pending_objects: Objects<R>,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum OnChain {
+    No,
+    Yes,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum Immutable {
+    No,
+    Yes,
+}
+
+#[derive(Debug)]
+pub enum MockError {
+    DuplicatedNullifier,
+    DuplicatedNote,
+}
+
+impl fmt::Display for MockError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for MockError {}
+
+impl<R: Rng + SeedableRng> MockChain<R> {
+    // CONSTRUCTORS
+    // ----------------------------------------------------------------------------------------
+
+    pub fn new(mut rng: R) -> Self {
+        let account_rng = R::from_rng(&mut rng).expect("rng seeding failed");
+        let account_id_builder = AccountIdBuilder::new(account_rng);
+        Self {
+            chain: ChainMmr::default(),
+            blocks: vec![],
+            nullifiers: TieredSmt::default(),
+            accounts: SimpleSmt::new(64).expect("depth too big for SimpleSmt"),
+            rng,
+            account_id_builder,
+            objects: Objects::new(),
+            pending_objects: Objects::new(),
+        }
+    }
+
+    // BUILDERS
+    // ----------------------------------------------------------------------------------------
+
+    /// Creates an [Account] and add to the list of pending objects.
+    pub fn build_account<C, S, A>(
+        &mut self,
+        code: C,
+        storage: S,
+        assets: A,
+        immutable: Immutable,
+        on_chain: OnChain,
+    ) -> AccountId
+    where
+        C: AsRef<str>,
+        S: IntoIterator<Item = StorageItem>,
+        A: IntoIterator<Item = Asset>,
+    {
+        let account_type = match immutable {
+            Immutable::Yes => AccountType::RegularAccountImmutableCode,
+            Immutable::No => AccountType::RegularAccountUpdatableCode,
+        };
+
+        let on_chain_bool = match on_chain {
+            OnChain::Yes => true,
+            OnChain::No => false,
+        };
+
+        let storage = AccountStorageBuilder::new().add_items(storage).build();
+
+        let (seed, _) = accountid_build_details(
+            &mut self.rng,
+            code.as_ref(),
+            account_type,
+            on_chain_bool,
+            storage.root(),
+        )
+        .unwrap();
+
+        let rng = R::from_rng(&mut self.rng).expect("rng seeding failed");
+        let account = AccountBuilder::new(rng)
+            .add_assets(assets)
+            .account_type(account_type)
+            .on_chain(on_chain == OnChain::Yes)
+            .code(code)
+            .with_seed_and_storage(seed, storage)
+            .unwrap();
+        let account_id = account.id();
+        self.pending_objects.accounts.push((account, seed));
+        account_id
+    }
+
+    /// Creates an [Account] using `seed` and add to the list of pending objects.
+    pub fn build_account_with_seed<C, S, A>(
+        &mut self,
+        seed: Word,
+        code: C,
+        storage: S,
+        assets: A,
+        immutable: Immutable,
+        on_chain: OnChain,
+    ) -> AccountId
+    where
+        C: AsRef<str>,
+        S: IntoIterator<Item = StorageItem>,
+        A: IntoIterator<Item = Asset>,
+    {
+        let account_type = match immutable {
+            Immutable::Yes => AccountType::RegularAccountImmutableCode,
+            Immutable::No => AccountType::RegularAccountUpdatableCode,
+        };
+
+        let rng = R::from_rng(&mut self.rng).expect("rng seeding failed");
+        let account = AccountBuilder::new(rng)
+            .add_storage_items(storage)
+            .add_assets(assets)
+            .account_type(account_type)
+            .on_chain(on_chain == OnChain::Yes)
+            .code(code)
+            .with_seed(seed)
+            .unwrap();
+        let account_id = account.id();
+        self.pending_objects.accounts.push((account, seed));
+        account_id
+    }
+
+    pub fn build_basic_wallet(&mut self) -> AccountId {
+        let account_type = AccountType::RegularAccountUpdatableCode;
+        let on_chain = true;
+        let storage = AccountStorageBuilder::new().build();
+        let (seed, _) = accountid_build_details(
+            &mut self.rng,
+            DEFAULT_ACCOUNT_CODE,
+            account_type,
+            on_chain,
+            storage.root(),
+        )
+        .unwrap();
+        let rng = R::from_rng(&mut self.rng).expect("rng seeding failed");
+        let account = AccountBuilder::new(rng)
+            .account_type(account_type)
+            .on_chain(on_chain)
+            .code(DEFAULT_ACCOUNT_CODE)
+            .build()
+            .unwrap();
+        let account_id = account.id();
+        self.pending_objects.accounts.push((account, seed));
+        account_id
+    }
+
+    /// Creates a [AccountId] with type [AccountType::FungibleFaucet] and add to the list of
+    /// pending objects.
+    pub fn build_fungible_faucet<C: AsRef<str>>(
+        &mut self,
+        on_chain: OnChain,
+        code: C,
+        storage_root: Digest,
+    ) -> AccountId {
+        let faucet_id = self
+            .account_id_builder
+            .account_type(AccountType::FungibleFaucet)
+            .on_chain(on_chain == OnChain::Yes)
+            .code(code)
+            .storage_root(storage_root)
+            .build()
+            .unwrap();
+        let builder = FungibleAssetBuilder::new(faucet_id)
+            .expect("builder was not configured to create fungible faucets");
+        self.pending_objects.fungible_faucets.push((faucet_id, builder));
+        faucet_id
+    }
+
+    /// Creates a [AccountId] with type [AccountType::FungibleFaucet] and add to the list of
+    /// pending objects.
+    pub fn build_fungible_faucet_with_seed<C: AsRef<str>>(
+        &mut self,
+        seed: Word,
+        on_chain: OnChain,
+        code: C,
+        storage_root: Digest,
+    ) -> AccountId {
+        let faucet_id = self
+            .account_id_builder
+            .account_type(AccountType::FungibleFaucet)
+            .on_chain(on_chain == OnChain::Yes)
+            .code(code)
+            .storage_root(storage_root)
+            .with_seed(seed)
+            .unwrap();
+        let builder = FungibleAssetBuilder::new(faucet_id)
+            .expect("builder was not configured to create fungible faucets");
+        self.pending_objects.fungible_faucets.push((faucet_id, builder));
+        faucet_id
+    }
+
+    /// Creates a [AccountId] with type [AccountType::NonFungibleFaucet] and add to the list of
+    /// pending objects.
+    pub fn build_nonfungible_faucet<C: AsRef<str>>(
+        &mut self,
+        on_chain: OnChain,
+        code: C,
+        storage_root: Digest,
+    ) -> AccountId {
+        let faucet_id = self
+            .account_id_builder
+            .account_type(AccountType::NonFungibleFaucet)
+            .on_chain(on_chain == OnChain::Yes)
+            .code(code)
+            .storage_root(storage_root)
+            .build()
+            .unwrap();
+        let rng = R::from_rng(&mut self.rng).expect("rng seeding failed");
+        let builder = NonFungibleAssetBuilder::new(faucet_id, rng)
+            .expect("builder was not configured to build nonfungible faucets");
+        self.pending_objects.nonfungible_faucets.push((faucet_id, builder));
+        faucet_id
+    }
+
+    /// Creates a [AccountId] with type [AccountType::NonFungibleFaucet] and add to the list of
+    /// pending objects.
+    pub fn build_nonfungible_faucet_with_seed<C: AsRef<str>>(
+        &mut self,
+        seed: Word,
+        on_chain: OnChain,
+        code: C,
+        storage_root: Digest,
+    ) -> AccountId {
+        let faucet_id = self
+            .account_id_builder
+            .account_type(AccountType::NonFungibleFaucet)
+            .on_chain(on_chain == OnChain::Yes)
+            .code(code)
+            .storage_root(storage_root)
+            .with_seed(seed)
+            .unwrap();
+        let rng = R::from_rng(&mut self.rng).expect("rng seeding failed");
+        let builder = NonFungibleAssetBuilder::new(faucet_id, rng)
+            .expect("builder was not configured to build nonfungible faucets");
+        self.pending_objects.nonfungible_faucets.push((faucet_id, builder));
+        faucet_id
+    }
+
+    /// Creates [FungibleAsset] from the fungible faucet at position `faucet_pos`.
+    pub fn build_fungible_asset(&mut self, faucet_pos: usize, amount: u64) -> Asset {
+        self.objects.fungible_faucets[faucet_pos]
+            .1
+            .amount(amount)
+            .unwrap()
+            .build()
+            .map(|v| v.into())
+            .unwrap()
+    }
+
+    /// Creates [NonFungibleAsset] from the nonfungible faucet at position `faucet_pos`.
+    pub fn build_nonfungible_asset(&mut self, faucet_pos: usize) -> Asset {
+        self.objects.nonfungible_faucets[faucet_pos]
+            .1
+            .build()
+            .map(|v| v.into())
+            .unwrap()
+    }
+
+    fn check_nullifier_unknown(&self, nullifier: Digest) {
+        assert!(self.pending_objects.nullifiers.iter().any(|e| *e == nullifier));
+        assert!(self.nullifiers.get_value(nullifier) != TieredSmt::EMPTY_VALUE)
+    }
+
+    // MODIFIERS
+    // ----------------------------------------------------------------------------------------
+
+    /// Creates the next block.
+    ///
+    /// This will also make all the objects currently pending available for use.
+    pub fn seal_block(&mut self) -> BlockHeader {
+        let block_num: u64 = self.blocks.len().try_into().expect("usize to u64 failed");
+        let block_num: Felt = block_num.into();
+
+        for (account, _seed) in self.pending_objects.accounts.iter() {
+            let id: Felt = account.id().into();
+            self.accounts.update_leaf(id.as_int(), account.hash().into()).unwrap();
+        }
+        for (account, _seed) in self.objects.accounts.iter() {
+            let id: Felt = account.id().into();
+            self.accounts.update_leaf(id.as_int(), account.hash().into()).unwrap();
+        }
+
+        // TODO:
+        // - resetting the nullifier tree once defined at the protocol level.
+        // - insering only nullifier from transactions included in the batches, once the batch
+        // kernel has been implemented.
+        for nullifier in self.pending_objects.nullifiers.iter() {
+            self.nullifiers
+                .insert(*nullifier, [block_num, Felt::ZERO, Felt::ZERO, Felt::ZERO]);
+        }
+        let notes = self.pending_objects.build_notes_tree();
+
+        let previous = self.blocks.last();
+        let peaks = self.chain.mmr().accumulator();
+        let chain_root: Digest = peaks.hash_peaks().into();
+        let account_root = self.accounts.root();
+        let prev_hash = previous.map_or(Digest::default(), |header| header.hash());
+        let nullifier_root = self.nullifiers.root();
+        let note_root = notes.root();
+        let version = Felt::ZERO;
+        let timestamp =
+            previous.map_or(TIMESTAMP_START, |header| header.timestamp() + TIMESTAMP_STEP);
+
+        // TODO: Set batch_root and proof_hash to the correct values once the kernel is
+        // available.
+        let batch_root = Digest::default();
+        let proof_hash = Digest::default();
+
+        let header = BlockHeader::new(
+            prev_hash,
+            block_num,
+            chain_root,
+            account_root,
+            nullifier_root,
+            note_root,
+            batch_root,
+            proof_hash,
+            version,
+            timestamp,
+        );
+
+        self.blocks.push(header);
+        self.chain.mmr_mut().add(header.hash());
+        self.objects.update_with(&mut self.pending_objects, header, &notes);
+
+        header
+    }
+
+    /// Mark a [Note] as produced by inserting into the block.
+    pub fn add_note(&mut self, note: Note) -> Result<(), MockError> {
+        if self.pending_objects.notes.iter().any(|e| e.hash() == note.hash()) {
+            return Err(MockError::DuplicatedNote);
+        }
+
+        // The check below works because the notes can not be added directly to the
+        // [BlockHeader], so we don't have to iterate over the known headers and check for
+        // inclusion proofs.
+        if self.objects.notes.iter().any(|e| e.hash() == note.hash()) {
+            return Err(MockError::DuplicatedNote);
+        }
+
+        self.check_nullifier_unknown(note.nullifier());
+        self.pending_objects.notes.push(note);
+        Ok(())
+    }
+
+    /// Mark a [Note] as consumed by inserting its nullifier into the block.
+    pub fn add_nullifier(&mut self, nullifier: Digest) -> Result<(), MockError> {
+        self.check_nullifier_unknown(nullifier);
+        self.pending_objects.nullifiers.push(nullifier);
+        Ok(())
+    }
+
+    /// Add a known [Account] to the mock chain.
+    pub fn add_account(&mut self, account: Account, seed: Word) {
+        assert!(
+            !self.pending_objects.accounts.iter().any(|(a, _)| a.id() == account.id()),
+            "Found duplicated AccountId"
+        );
+        self.pending_objects.accounts.push((account, seed));
+    }
+
+    // ACESSORS
+    // ----------------------------------------------------------------------------------------
+
+    /// Get a reference to the latest [ChainMmr].
+    pub fn chain(&self) -> &ChainMmr {
+        &self.chain
+    }
+
+    /// Get a reference to [BlockHeader] with `block_number`.
+    pub fn blockheader(&self, block_number: usize) -> &BlockHeader {
+        &self.blocks[block_number]
+    }
+
+    /// Get a reference to the nullifier tree.
+    pub fn nullifiers(&self) -> &TieredSmt {
+        &self.nullifiers
+    }
+
+    /// Get the [AccountId] of the nth fungible faucet.
+    pub fn fungible(&self, faucet_pos: usize) -> AccountId {
+        self.objects.fungible_faucets[faucet_pos].0
+    }
+
+    /// Get the [AccountId] of the nth nonfungible faucet.
+    pub fn nonfungible(&self, faucet_pos: usize) -> AccountId {
+        self.objects.nonfungible_faucets[faucet_pos].0
+    }
+
+    /// Get a mutable reference to nth [Account].
+    pub fn account_mut(&mut self, pos: usize) -> &mut Account {
+        &mut self.objects.accounts[pos].0
+    }
+
+    /// Get the [Account]'s corresponding seed.
+    pub fn account_seed(&mut self, pos: usize) -> Word {
+        self.objects.accounts[pos].1
+    }
+}
+
+// SERIALIZATION
+// --------------------------------------------------------------------------------------------
+#[cfg(feature = "serde")]
+use std::{
+    fs::File,
+    io::{self, Read, Write},
+    path::Path,
+};
+
+#[cfg(feature = "serde")]
+impl<R: serde::Serialize> MockChain<R> {
+    pub fn to_file<T: AsRef<Path>>(self, path: T) -> io::Result<()> {
+        let encoded = postcard::to_allocvec(&self).unwrap();
+        File::create(path)?.write_all(&encoded)?;
+        Ok(())
+    }
+}
+
+#[cfg(feature = "serde")]
+pub fn from_file<T, P>(path: P) -> io::Result<MockChain<T>>
+where
+    P: AsRef<Path>,
+    MockChain<T>: serde::de::DeserializeOwned,
+{
+    let mut data = Vec::new();
+    File::open(path)?.read_to_end(&mut data)?;
+    let data: MockChain<T> = postcard::from_bytes(&data).unwrap();
+    Ok(data)
+}
